+++ src/neptune_tensorboard/integration/tensorflow_integration.py
@@ -13,12 +13,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+import contextlib
 import warnings
-from functools import wraps
 from importlib.util import find_spec
 
 import tensorflow as tf
 from neptune.types import File
+
+from neptune_tensorboard.integration.utils import register_pre_hook
 
 IS_GRAPHLIB_AVAILABLE = find_spec("tfgraphviz")
 if IS_GRAPHLIB_AVAILABLE:
@@ -26,14 +28,14 @@ if IS_GRAPHLIB_AVAILABLE:
 
 _integrated_with_tensorflow = False
 
-__all__ = ["patch_tensorflow"]
+__all__ = ["patch_tensorflow", "NeptuneTensorflowTracker"]
 
 
 def patch_tensorflow(run, base_namespace):
     global _integrated_with_tensorflow
 
     if not _integrated_with_tensorflow:
-        patch_tensorflow_2x(run, base_namespace)
+        NeptuneTensorflowTracker(run, base_namespace)
         _integrated_with_tensorflow = True
 
 
@@ -67,25 +69,28 @@ def track_graph(graph_data, run=None, base_namespace=None):
         warnings.warn("Skipping model visualization because no tfgraphviz installation was found.")
 
 
-def register_pre_hook(original, neptune_hook, run, base_namespace):
-    @wraps(original)
-    def wrapper(*args, **kwargs):
-        neptune_hook(*args, **kwargs, run=run, base_namespace=base_namespace)
-        return original(*args, **kwargs)
+class NeptuneTensorflowTracker(contextlib.AbstractContextManager):
+    def __init__(self, run, base_namespace):
+        self.org_scalar = tf.summary.scalar
+        self.org_image = tf.summary.image
+        self.org_text = tf.summary.text
+        self.org_graph = tf.summary.graph
 
-    return wrapper
+        tf.summary.scalar = register_pre_hook(
+            original=tf.summary.scalar, neptune_hook=track_scalar, run=run, base_namespace=base_namespace
+        )
+        tf.summary.image = register_pre_hook(
+            original=tf.summary.image, neptune_hook=track_image, run=run, base_namespace=base_namespace
+        )
+        tf.summary.text = register_pre_hook(
+            original=tf.summary.text, neptune_hook=track_text, run=run, base_namespace=base_namespace
+        )
+        tf.summary.graph = register_pre_hook(
+            original=tf.summary.graph, neptune_hook=track_graph, run=run, base_namespace=base_namespace
+        )
 
-
-def patch_tensorflow_2x(run, base_namespace):
-    tf.summary.scalar = register_pre_hook(
-        original=tf.summary.scalar, neptune_hook=track_scalar, run=run, base_namespace=base_namespace
-    )
-    tf.summary.image = register_pre_hook(
-        original=tf.summary.image, neptune_hook=track_image, run=run, base_namespace=base_namespace
-    )
-    tf.summary.text = register_pre_hook(
-        original=tf.summary.text, neptune_hook=track_text, run=run, base_namespace=base_namespace
-    )
-    tf.summary.graph = register_pre_hook(
-        original=tf.summary.graph, neptune_hook=track_graph, run=run, base_namespace=base_namespace
-    )
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        tf.summary.scalar = self.org_scalar
+        tf.summary.image = self.org_image
+        tf.summary.text = self.org_text
+        tf.summary.graph = self.org_graph